chore: convert to mocks structure
