+++ docs/app/views/pages/sandbox.html.erb
@@ -1,245 +1,7 @@
-
-<%
-
-# mock_data_as = "panel_rows"
-mock_data_as = "description"
-
-# potential data structure
-data = {
-  avatar_image: "avatar/court.png",
-  email: "court@kajabi.com",
-  initials: "CM",
-  name: "Court McFadzean",
-  user_stats: [
-    [
-      {
-        title: "Lifetime value",
-        data: "$230",
-      },
-      {
-        title: "Total offers",
-        data: "3",
-        link: "#offers"
-      },
-      {
-        title: "Total products",
-        data: "6",
-        link: "#products"
-      },
-    ],
-    [
-      {
-        title: "Last sign-in",
-        data: "Aug. 12, 2021",
-      },
-      {
-        title: "Total sign-in",
-        data: "32",
-      },
-    ],
-    [
-      {
-        title: "Contact status",
-        data: sage_component(SageLabel, {
-          value: "Member",
-          color: "info",
-          icon: "check-circle-filled"
-        }),
-      },
-      {
-        title: "Marketing status",
-        data: "Subscribed",
-      },
-      {
-        title: "Contact added",
-        data: "Jan. 23, 2021",
-      },
-      {
-        title: "Opt-in status",
-        data: "Unconfirmed",
-      },
-      {
-        title: "Affiliate referral",
-        data: "Abe Lincoln",
-      },
-    ]
-  ],
-  tags: ["2021", "Hello", "Summer", "Holiday", "Welcome", "Tag", "Tuesday"]
-}
-
-# configs for profile buttons
-button_bar_items = [
-  {
-    href: "#edit-contact",
-    icon: "pen",
-    label: "Edit contact",
-  },
-  {
-    href: "#reset-password",
-    icon: "ban", # TODO: Need reset password icon
-    label: "Reset password",
-  },
-  {
-    href: "#add-tag",
-    icon: "tag",
-    label: "Add tag",
-  },
-  {
-    href: "#grant-offer",
-    icon: "grant-offer",
-    label: "Grant offer",
-  },
-]
-
-# configs for more options dropdown
-more_options_items = [
-  {
-    value: "Change password",
-    attributes: { href: "#change-password" },
-  },
-  {
-    value: "Add a note",
-    attributes: { href: "#add-a-note" },
-  },
-  {
-    value: "Mute",
-    attributes: { href: "#mute" },
-  },
-  {
-    value: "Hide",
-    attributes: { href: "#hide" },
-  },
-  {
-    value: "Delete",
-    attributes: { href: "#delete" },
-    style: "danger",
-    modifiers: ["border-before"],
-  }
-]
-
-# ------------------------------------------------------------------------------------------
-
-%>
-
 <%= content_for :heading do %>
 <%= md("
 # Sandbox
 
 Use this scratchpad to test out elements and components in context. Just be sure not to commit your changes. ❤️
 ") %>
-<% end %>
-<%= sage_component SageContainer, { size: "sm" } do %>
-  
-  <%# TODO: Need a drawer or modal to house this content rather than Panel %>
-  <%= sage_component SagePanel, {} do %>
-
-    <%# --- Header --- %>
-    <%= sage_component SagePanelRow, { grid_template: "te" } do %>
-      <%= sage_component SageButton, {
-        attributes: { href: "#TODO-open-full-profile" },
-        value: "Open full profile",
-        subtle: true,
-        small: true,
-        style: "secondary",
-        icon: {
-          name: "expand",
-          style: "left",
-        },
-      } %>
-      <%= sage_component SageButton, {
-        attributes: { href: "#TODO-close-preview" },
-        icon: {
-          name: "remove",
-          style: "only",
-        },
-        subtle: true,
-        style: "secondary",
-        value: "Close preview",
-      } %>
-    <% end %>
-    <%# /-- Header --- %>
-
-    <%# --- Profile --- %>
-    <%= sage_component SagePanelBlock, { css_classes: SageClassnames::TYPE_ALIGN_CENTER } do %>
-      <%= sage_component SageAvatar, {
-        centered: true,
-        image: {
-          src: data[:avatar_image],
-          alt: data[:name],
-        },
-        initials: data[:initials],
-        size: "68px",
-        spacer: { bottom: "sm" },
-        # TODO: Add `badge` property
-      } %>
-      <h3 class="<%= SageClassnames::TYPE::HEADING_4 %> <%= SageClassnames::TYPE_COLORS::CHARCOAL_500 %>">
-        <%= data[:name] %>
-      </h3>
-      <%= sage_component SageCopyButton, {
-        value: data[:email],
-        borderless: true,
-      } %>
-    <% end %>
-    <%= sage_component SageButtonGroup, {
-      gap: :sm,
-      align: "center"
-    } do %>
-      <% button_bar_items.each do | btn | %>
-        <%= sage_component SageButton, {
-          value: btn[:label],
-          style: "secondary",
-          raised: false,
-          icon: {
-            style: "only",
-            name: btn[:icon],
-          },
-          html_attributes: {
-            "data-js-tooltip": btn[:label],
-            href: btn[:href],
-          },
-        } %>
-      <% end %>
-      <%= sage_component SageDropdown, { items: more_options_items } do %>
-        <%= sage_component SageButton, {
-          value: "More",
-          style: "secondary",
-          raised: false,
-          icon: {
-            style: "only",
-            name: "dot-menu-horizontal",
-          },
-          html_attributes: {
-            "data-js-tooltip": "More",
-          },
-        } %>
-      <% end %>
-    <% end %>
-    <%# /-- Profile --- %>
-
-    <%# --- Data --- %>
-    <% data[:user_stats].each do | data_group | %>
-      <%= sage_component SageDescription, {
-        allcaps_titles: false,
-        inline_spread: true,
-        items: data_group,
-      } %>
-      <%= sage_component SagePanelDivider, {} %>
-    <% end %>
-
-    <%= sage_component SageLabelGroup, { gap: :xs } do %>
-      <% data[:tags].each do | tag | %>
-        <%= sage_component SageLabel, { color: "draft", value: tag } %>
-      <% end %>
-      <%= sage_component SageButton, {
-        attributes: {
-          href: "#view-more",
-        },
-        style: "primary",
-        subtle: true,
-        value: "View all",
-      } %>
-    <% end %>
-    <%# /-- Data --- %>
-
-  <% end %>
 <% end %>
