+++ packages/sage-react/lib/mocks/contact-profile/components/index.js
@@ -1,0 +1,3 @@
+export { Header } from './Header';
+export { Profile } from './Profile';
+export { Data } from './Data';
